Add files via upload
About: Prediction of Loan Eligibility
 Dream Housing Finance company deals in all kinds of home loans. They have a presence across all urban, semi-urban and rural areas. The customer first applies for a home loan and after that, the company validates the customer eligibility for the loan. The company wants to automate the loan eligibility process(real-time) based on customer detail provided while filling our online application forms. These details are Gender, Marital Status, Education, number of Dependents, income, loan amount, credit History, and others. To automate this process, they have provided a dataset to identify the customer segments that are eligible for loan amounts so that they can specifically target these customers.

Data Description: The source of the data is kaggle.com. This loan data contains 614 rows and 13 columns. In this data there are various eligibility criteria are given like, Gender, Education, Dependents, Loan Amount, etc.
Objective:1.To predict whether the applicants are eligible for loan or not, according to the eligibility criteria. 2.Build various ML models that can predict the loan approval.

Conclusion: According this I have concluded that the Logistic Regression and Support Vector Machine Classification these algorithms of machine learning are good fit for data as compared to other algorithms.